dbt_mcp_hackathon_project/requirements.txt
# Core dependencies
streamlit>=1.28.0
dbt-core>=1.6.0
dbt-duckdb>=1.6.0

# MCP server dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0

# Data processing and visualization
pandas>=2.0.0
networkx>=3.2
plotly>=5.17.0

# AI integration
openai>=1.3.0
requests>=2.31.0

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0
typer>=0.9.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0